flask_app/requirements.txt
Babel==1.3
blinker==1.3
coverage==3.7.1
decorator==3.4.0
Flask==0.10.1
Flask-Babel==0.9
Flask-HTTPAuth==2.4.0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-OpenID==1.2.4
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.11
flipflop==1.0
guess-language==0.2
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
passlib==1.6.2
pbr==0.10.7
Pillow==2.7.0
psycopg2==2.6
python-openid==2.2.5
pytz==2014.10
requests==2.5.1
six==1.9.0
speaklater==1.3
SQLAlchemy==0.9.8
sqlalchemy-migrate==0.9.4
sqlparse==0.1.14
stripe==1.20.2
Tempita==0.5.2
Werkzeug==0.10
Whoosh==2.6.0
WTForms==2.0.2
